break[array]: move cast traits to exprs
