+++ encodings/fastlanes/public-api.lock
@@ -101,10 +101,10 @@ impl vortex_array::arrays::slice::SliceKernel for vortex_fastlanes::BitPackedVTa
 pub fn vortex_fastlanes::BitPackedVTable::slice(array: &vortex_fastlanes::BitPackedArray, range: core::ops::range::Range<usize>, _ctx: &mut vortex_array::executor::ExecutionCtx) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::compute::between::BetweenKernel for vortex_fastlanes::BitPackedVTable
 pub fn vortex_fastlanes::BitPackedVTable::between(&self, array: &vortex_fastlanes::BitPackedArray, lower: &dyn vortex_array::array::Array, upper: &dyn vortex_array::array::Array, options: &vortex_array::compute::between::BetweenOptions) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
-impl vortex_array::compute::cast::CastReduce for vortex_fastlanes::BitPackedVTable
-pub fn vortex_fastlanes::BitPackedVTable::cast(array: &vortex_fastlanes::BitPackedArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::compute::is_constant::IsConstantKernel for vortex_fastlanes::BitPackedVTable
 pub fn vortex_fastlanes::BitPackedVTable::is_constant(&self, array: &vortex_fastlanes::BitPackedArray, opts: &vortex_array::compute::is_constant::IsConstantOpts) -> vortex_error::VortexResult<core::option::Option<bool>>
+impl vortex_array::expr::exprs::cast::kernel::CastReduce for vortex_fastlanes::BitPackedVTable
+pub fn vortex_fastlanes::BitPackedVTable::cast(array: &vortex_fastlanes::BitPackedArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::vtable::VTable for vortex_fastlanes::BitPackedVTable
 pub type vortex_fastlanes::BitPackedVTable::Array = vortex_fastlanes::BitPackedArray
 pub type vortex_fastlanes::BitPackedVTable::ArrayVTable = vortex_fastlanes::BitPackedVTable
@@ -169,7 +169,7 @@ impl core::fmt::Debug for vortex_fastlanes::DeltaVTable
 pub fn vortex_fastlanes::DeltaVTable::fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result
 impl vortex_array::arrays::slice::SliceReduce for vortex_fastlanes::DeltaVTable
 pub fn vortex_fastlanes::DeltaVTable::slice(array: &Self::Array, range: core::ops::range::Range<usize>) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
-impl vortex_array::compute::cast::CastReduce for vortex_fastlanes::DeltaVTable
+impl vortex_array::expr::exprs::cast::kernel::CastReduce for vortex_fastlanes::DeltaVTable
 pub fn vortex_fastlanes::DeltaVTable::cast(array: &vortex_fastlanes::DeltaArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::vtable::VTable for vortex_fastlanes::DeltaVTable
 pub type vortex_fastlanes::DeltaVTable::Array = vortex_fastlanes::DeltaArray
@@ -231,8 +231,6 @@ impl vortex_array::arrays::filter::kernel::FilterReduce for vortex_fastlanes::Fo
 pub fn vortex_fastlanes::FoRVTable::filter(array: &vortex_fastlanes::FoRArray, mask: &vortex_mask::Mask) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::arrays::slice::SliceReduce for vortex_fastlanes::FoRVTable
 pub fn vortex_fastlanes::FoRVTable::slice(array: &Self::Array, range: core::ops::range::Range<usize>) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
-impl vortex_array::compute::cast::CastReduce for vortex_fastlanes::FoRVTable
-pub fn vortex_fastlanes::FoRVTable::cast(array: &vortex_fastlanes::FoRArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::compute::compare::CompareKernel for vortex_fastlanes::FoRVTable
 pub fn vortex_fastlanes::FoRVTable::compare(&self, lhs: &vortex_fastlanes::FoRArray, rhs: &dyn vortex_array::array::Array, operator: vortex_array::compute::compare::Operator) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::compute::is_constant::IsConstantKernel for vortex_fastlanes::FoRVTable
@@ -240,6 +238,8 @@ pub fn vortex_fastlanes::FoRVTable::is_constant(&self, array: &vortex_fastlanes:
 impl vortex_array::compute::is_sorted::IsSortedKernel for vortex_fastlanes::FoRVTable
 pub fn vortex_fastlanes::FoRVTable::is_sorted(&self, array: &vortex_fastlanes::FoRArray) -> vortex_error::VortexResult<core::option::Option<bool>>
 pub fn vortex_fastlanes::FoRVTable::is_strict_sorted(&self, array: &vortex_fastlanes::FoRArray) -> vortex_error::VortexResult<core::option::Option<bool>>
+impl vortex_array::expr::exprs::cast::kernel::CastReduce for vortex_fastlanes::FoRVTable
+pub fn vortex_fastlanes::FoRVTable::cast(array: &vortex_fastlanes::FoRArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::vtable::VTable for vortex_fastlanes::FoRVTable
 pub type vortex_fastlanes::FoRVTable::Array = vortex_fastlanes::FoRArray
 pub type vortex_fastlanes::FoRVTable::ArrayVTable = vortex_fastlanes::FoRVTable
@@ -306,7 +306,7 @@ impl core::fmt::Debug for vortex_fastlanes::RLEVTable
 pub fn vortex_fastlanes::RLEVTable::fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result
 impl vortex_array::arrays::slice::SliceKernel for vortex_fastlanes::RLEVTable
 pub fn vortex_fastlanes::RLEVTable::slice(array: &vortex_fastlanes::RLEArray, range: core::ops::range::Range<usize>, _ctx: &mut vortex_array::executor::ExecutionCtx) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
-impl vortex_array::compute::cast::CastReduce for vortex_fastlanes::RLEVTable
+impl vortex_array::expr::exprs::cast::kernel::CastReduce for vortex_fastlanes::RLEVTable
 pub fn vortex_fastlanes::RLEVTable::cast(array: &vortex_fastlanes::RLEArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::vtable::VTable for vortex_fastlanes::RLEVTable
 pub type vortex_fastlanes::RLEVTable::Array = vortex_fastlanes::RLEArray
