+++ encodings/fsst/public-api.lock
@@ -51,10 +51,10 @@ impl vortex_array::arrays::filter::kernel::FilterKernel for vortex_fsst::FSSTVTa
 pub fn vortex_fsst::FSSTVTable::filter(array: &vortex_fsst::FSSTArray, mask: &vortex_mask::Mask, ctx: &mut vortex_array::executor::ExecutionCtx) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::arrays::slice::SliceReduce for vortex_fsst::FSSTVTable
 pub fn vortex_fsst::FSSTVTable::slice(array: &Self::Array, range: core::ops::range::Range<usize>) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
-impl vortex_array::compute::cast::CastReduce for vortex_fsst::FSSTVTable
-pub fn vortex_fsst::FSSTVTable::cast(array: &vortex_fsst::FSSTArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::compute::compare::CompareKernel for vortex_fsst::FSSTVTable
 pub fn vortex_fsst::FSSTVTable::compare(&self, lhs: &vortex_fsst::FSSTArray, rhs: &dyn vortex_array::array::Array, operator: vortex_array::compute::compare::Operator) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
+impl vortex_array::expr::exprs::cast::kernel::CastReduce for vortex_fsst::FSSTVTable
+pub fn vortex_fsst::FSSTVTable::cast(array: &vortex_fsst::FSSTArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::vtable::VTable for vortex_fsst::FSSTVTable
 pub type vortex_fsst::FSSTVTable::Array = vortex_fsst::FSSTArray
 pub type vortex_fsst::FSSTVTable::ArrayVTable = vortex_fsst::FSSTVTable
