+++ encodings/runend/public-api.lock
@@ -58,8 +58,6 @@ impl vortex_array::arrays::dict::take::TakeExecute for vortex_runend::RunEndVTab
 pub fn vortex_runend::RunEndVTable::take(array: &vortex_runend::RunEndArray, indices: &dyn vortex_array::array::Array, _ctx: &mut vortex_array::executor::ExecutionCtx) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::arrays::filter::kernel::FilterKernel for vortex_runend::RunEndVTable
 pub fn vortex_runend::RunEndVTable::filter(array: &vortex_runend::RunEndArray, mask: &vortex_mask::Mask, _ctx: &mut vortex_array::executor::ExecutionCtx) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
-impl vortex_array::compute::cast::CastReduce for vortex_runend::RunEndVTable
-pub fn vortex_runend::RunEndVTable::cast(array: &vortex_runend::RunEndArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::compute::compare::CompareKernel for vortex_runend::RunEndVTable
 pub fn vortex_runend::RunEndVTable::compare(&self, lhs: &vortex_runend::RunEndArray, rhs: &dyn vortex_array::array::Array, operator: vortex_array::compute::compare::Operator) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::compute::is_constant::IsConstantKernel for vortex_runend::RunEndVTable
@@ -69,6 +67,8 @@ pub fn vortex_runend::RunEndVTable::is_sorted(&self, array: &vortex_runend::RunE
 pub fn vortex_runend::RunEndVTable::is_strict_sorted(&self, array: &vortex_runend::RunEndArray) -> vortex_error::VortexResult<core::option::Option<bool>>
 impl vortex_array::compute::min_max::MinMaxKernel for vortex_runend::RunEndVTable
 pub fn vortex_runend::RunEndVTable::min_max(&self, array: &vortex_runend::RunEndArray) -> vortex_error::VortexResult<core::option::Option<vortex_array::compute::min_max::MinMaxResult>>
+impl vortex_array::expr::exprs::cast::kernel::CastReduce for vortex_runend::RunEndVTable
+pub fn vortex_runend::RunEndVTable::cast(array: &vortex_runend::RunEndArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::expr::exprs::fill_null::kernel::FillNullReduce for vortex_runend::RunEndVTable
 pub fn vortex_runend::RunEndVTable::fill_null(array: &vortex_runend::RunEndArray, fill_value: &vortex_scalar::scalar::Scalar) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::vtable::VTable for vortex_runend::RunEndVTable
