+++ encodings/zigzag/public-api.lock
@@ -29,10 +29,10 @@ impl vortex_array::arrays::filter::kernel::FilterReduce for vortex_zigzag::ZigZa
 pub fn vortex_zigzag::ZigZagVTable::filter(array: &vortex_zigzag::ZigZagArray, mask: &vortex_mask::Mask) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::arrays::slice::SliceReduce for vortex_zigzag::ZigZagVTable
 pub fn vortex_zigzag::ZigZagVTable::slice(array: &Self::Array, range: core::ops::range::Range<usize>) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
-impl vortex_array::compute::cast::CastReduce for vortex_zigzag::ZigZagVTable
-pub fn vortex_zigzag::ZigZagVTable::cast(array: &vortex_zigzag::ZigZagArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::compute::mask::MaskKernel for vortex_zigzag::ZigZagVTable
 pub fn vortex_zigzag::ZigZagVTable::mask(&self, array: &vortex_zigzag::ZigZagArray, filter_mask: &vortex_mask::Mask) -> vortex_error::VortexResult<vortex_array::array::ArrayRef>
+impl vortex_array::expr::exprs::cast::kernel::CastReduce for vortex_zigzag::ZigZagVTable
+pub fn vortex_zigzag::ZigZagVTable::cast(array: &vortex_zigzag::ZigZagArray, dtype: &vortex_dtype::dtype::DType) -> vortex_error::VortexResult<core::option::Option<vortex_array::array::ArrayRef>>
 impl vortex_array::vtable::VTable for vortex_zigzag::ZigZagVTable
 pub type vortex_zigzag::ZigZagVTable::Array = vortex_zigzag::ZigZagArray
 pub type vortex_zigzag::ZigZagVTable::ArrayVTable = vortex_zigzag::ZigZagVTable
