+++ vortex-array/src/compute/mod.rs
@@ -17,7 +17,8 @@ use std::fmt::Formatter;
 use arcref::ArcRef;
 pub use between::*;
 pub use boolean::*;
-pub use cast::*;
+#[expect(deprecated)]
+pub use cast::cast;
 pub use compare::*;
 pub use fill_null::*;
 pub use filter::*;
@@ -46,6 +47,10 @@ pub use zip::*;
 use crate::Array;
 use crate::ArrayRef;
 use crate::builders::ArrayBuilder;
+pub use crate::expr::CastExecuteAdaptor;
+pub use crate::expr::CastKernel;
+pub use crate::expr::CastReduce;
+pub use crate::expr::CastReduceAdaptor;
 pub use crate::expr::FillNullExecuteAdaptor;
 pub use crate::expr::FillNullKernel;
 pub use crate::expr::FillNullReduce;
